+++ src/inspector/components/DescriptorItem/DescriptorItem.tsx
@@ -1,6 +1,6 @@
 import React from "react";
-import { IDescriptor, TSelectDescriptorOperation } from "../reducers/initialStateInspector";
 import "./DescriptorItem.css";
+import { IDescriptor, TSelectDescriptorOperation } from "../../model/types";
 
 export interface IDescriptorItemProps{
 	descriptor:IDescriptor
@@ -10,15 +10,11 @@ export interface IDescriptorItemDispatch {
 	onSelect:(uuid:string,operation:TSelectDescriptorOperation)=>void
 }
 
-export interface IDescriptorItemState{
-	
-}
-
 
 export type TDescriptorItem = IDescriptorItemProps & IDescriptorItemDispatch
-export type TDescriptorItemComponent = React.Component<TDescriptorItem, IDescriptorItemState>
+export type TDescriptorItemComponent = React.Component<TDescriptorItem>
 
-export class DescriptorItem extends React.Component<TDescriptorItem, IDescriptorItemState> { 
+export class DescriptorItem extends React.Component<TDescriptorItem> { 
 	constructor(props: TDescriptorItem) {
 		super(props);
 		
@@ -38,11 +34,28 @@ export class DescriptorItem extends React.Component<TDescriptorItem, IDescriptor
 		this.props.onSelect(this.props.descriptor.id, operation);
 	}
 
+	private generateItemName = () => {
+		const {descriptor:{originalReference,locked,pinned} } = this.props;
+		const slug = originalReference._target.map(r => {
+			if ("_ref" in r) {
+				return r._ref;				
+			}
+			if ("_property" in r) {
+				return r._property;				
+			}
+			return "n/a";
+		});
+		const lockedStr = locked ? " [L] " : "";
+		const pinnedStr = pinned ? " [P] " : "";
+		return slug.reverse().join(" / ") + lockedStr + pinnedStr;
+
+	}
+
 	public render():React.ReactNode{
 		return (
 			<div className={"DescriptorItem" + (this.props.descriptor.selected ? " selected" : "")} onClick={this.select}>
-				<span>{this.props.descriptor.id}</span>
-				<span className="time">{this.props.descriptor.endTime-this.props.descriptor.startTime}ms</span>
+				<span className="name">{this.generateItemName()}</span>
+				<span className="time">{this.props.descriptor.endTime-this.props.descriptor.startTime} ms</span>
 			</div>
 		);
 	}