+++ src/inspector/components/Tabs/TabList.tsx
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-empty-interface */
 import React from "react";
 import { TTabPanelComponent } from "./TabPanel";
 import "./TabList.css";