+++ src/store.tsx
@@ -2,8 +2,8 @@ import { createStore, applyMiddleware, combineReducers } from "redux";
 import { composeWithDevTools } from "redux-devtools-extension";
 import { inspectorReducer } from "./inspector/reducers";
 import { listenerReducer } from "./listener/reducers";
-import { IInspectorState } from "./inspector/reducers/initialStateInspector";
 import { IAppState } from "./listener/reducers/initialStateListener";
+import { IInspectorState } from "./inspector/model/types";
 
 /*
  * We're giving State interface to create store